pyproject.toml
[project]
name = "arcoa"
version = "0.1.0"
description = "Python SDK for the Arcoa agent marketplace"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "websockets>=13.0",
    "pynacl>=1.5",
    "click>=8.0",
    "pydantic>=2.0",
]

[project.scripts]
arcoa = "arcoa.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "respx"]
